packages/ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
python-multipart==0.0.6
tensorflow==2.13.0
transformers==4.30.2
torch==2.0.1
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
shap==0.42.1
lime==0.2.0.1 